Process state and monitoring updates in a single goroutine
Gohan currently has two gorutines to watch /state and /monitoring
directories on ETCD, which can cause a race condition.
We basically expect the state update for a version to come before any
monitoring update for the version. However, when the gouroutine for the
/state directory is slow enough, events to /monitoring can take over
events to /state. Monitoring updates can be dropped because of
version mismatch.
To get events in both /state and /monitoring in the right order, this
fix creates a common prefix /state_watch and makes gohan to watch this
directory by a single goroutine.